update: 更新文件 operator_config_cache.h

+++ services/sim/include/operator_config_cache.h
@@ -77,7 +77,7 @@ private:
     bool isUpdateImsCapFromChipDone_ = false;
     int32_t retryBatchInsertApnTimes_ = 0;
     static constexpr int32_t BATCH_INSERT_APN_RETRY_TIMES = 5;
-    static constexpr int32_t BATCH_INSERT_APN_RETRY_DEALY = 10*1000;
+    static constexpr int64_t BATCH_INSERT_APN_RETRY_DEALY = 10*1000;
     enum {
         RADIO_BATCH_INSERT_APN_RETRY = 0,
     };
